requirements.txt
streamlit>=1.30.0
opencv-python>=4.9.0.80
easyocr>=1.7.1
torch>=2.2.0
torchvision>=0.17.0
python-docx>=1.1.0
numpy>=1.26.0
Pillow>=10.0.0
